backend/requirements.txt
Django==4.1.5
pre-commit==2.21.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
Pillow==9.2.0
